coding standards fixes
